start separating CLI from library
