+++ decoder/nhw_decoder.h
@@ -1,0 +1,52 @@
+/***************************************************************************
+****************************************************************************
+*  NHW Image Codec                                                         *
+*  file: nhw_decoder.h                                                     *
+*  version: 0.1.6                                                          *
+*  last update: $ 18112022 nhw exp $                                       *
+*                                                                          *
+****************************************************************************
+****************************************************************************
+
+****************************************************************************
+*  remark: -simple codec                                                   *
+***************************************************************************/
+
+/* Copyright (C) 2007-2013 NHW Project
+   Written by Raphael Canut - nhwcodec_at_gmail.com */
+/*
+   Redistribution and use in source and binary forms, with or without
+   modification, are permitted provided that the following conditions
+   are met:
+
+   - Redistributions of source code must retain the above copyright
+   notice, this list of conditions and the following disclaimer.
+
+   - Redistributions in binary form must reproduce the above copyright
+   notice, this list of conditions and the following disclaimer in the
+   documentation and/or other materials provided with the distribution.
+
+   - Neither the name of NHW Codec, or NHW Project, nor the names of 
+   specific contributors, may be used to endorse or promote products 
+   derived from this software without specific prior written permission.
+
+   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+   ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+   LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+   A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+   OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+   EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+   PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+   PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+   LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+   NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+   SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+#ifndef _NHW_DECODER_H
+#define _NHW_DECODER_H
+
+
+#endif // _NHW_DECODER_H
+
+
